fix: controllers set to use service classes
