+++ backend/src/repos/index.js
@@ -1,6 +1,6 @@
-export * from './class/UserRepo';
-export * from './class/KingdomRepo';
-export * from './class/LocationRepo';
-export * from './class/ResourceRepo';
-export * from './class/BuildingRepo';
+export * from './UserRepo';
+export * from './KingdomRepo';
+export * from './LocationRepo';
+export * from './ResourceRepo';
+export * from './BuildingRepo';
 export * from './errorCodes';
