+++ backend/src/repos/kingdomRepo.js
@@ -1,32 +1,52 @@
-import {db} from '../data/connection';
-import { save } from './repoSave';
-import { paramsValidation } from './repoValidation';
+import QueryHandler from './QueryHandler';
 
-const insertQueries = {
-    kingdom : 'INSERT INTO kingdoms (name) VALUES(?)',
-    user_kingdom : 'INSERT INTO users_kingdoms (user_id,kingdom_id) VALUES(?,?)'
+export class KingdomRepo extends QueryHandler {
+
+    constructor(db,errorCodes) {
+        super(db,errorCodes);
+    };
+
+    async getById({kingdomId}) {
+        if (!kingdomId) throw new Error(this.errorCodes.missingKingdomId);
+        const query = this.validateQuery`
+        SELECT
+            kingdoms.id kingdomId,
+            kingdoms.name kingdomName,
+            users_kingdoms.user_id userId,
+            locations.code location
+        FROM kingdoms
+        LEFT JOIN users_kingdoms ON users_kingdoms.kingdom_id = kingdoms.id
+        LEFT JOIN locations ON locations.kingdom_id = kingdoms.id
+        WHERE kingdoms.id = ${kingdomId}
+        `;
+        const dbData = await (this.sendQuery(query));
+        if (dbData.length === 0) throw new Error(this.errorCodes.invalidKingdomId);
+        return dbData;
+    };
+
+    async add({kingdomName}) {
+        if (!kingdomName) throw new Error(this.errorCodes.missingKingdomName);
+        const query = this.validateQuery`INSERT INTO kingdoms (name) VALUES(${kingdomName})`;
+        return await (this.sendQuery(query));
+    };
+
+    async attachUser({kingdomId, userId}) {
+        if (!kingdomId) throw new Error(this.errorCodes.missingKingdomId);
+        if (!userId) throw new Error(this.errorCodes.missingUserId);
+        const query = this.validateQuery`INSERT INTO users_kingdoms (user_id,kingdom_id) VALUES(${userId},${kingdomId})`;
+        return await (this.sendQuery(query));
+    };
+
+    async get() {
+        const query = this.validateQuery`
+        SELECT
+            kingdoms.id kingdomId,
+            kingdoms.name kingdomName,
+            locations.code location
+        FROM kingdoms
+        LEFT JOIN locations ON locations.kingdom_id = kingdoms.id
+        `;
+        return await (this.sendQuery(query));
+    };
+
 };
-
-const selectQueries = {
-    KingdomNullLocation :
-    'SELECT kingdoms.id kingdomid, locations.id locationid FROM kingdoms LEFT JOIN locations ON locations.kingdom_id = kingdoms.id WHERE kingdoms.id = ? AND locations.id is null',
-    KingdomBaseData :
-    'SELECT users.id userid, kingdoms.name kingdomname FROM kingdoms RIGHT JOIN users_kingdoms ON users_kingdoms.kingdom_id = kingdoms.id RIGHT JOIN users ON users.id = users_kingdoms.user_id WHERE kingdom_id = ?',
-    KingdomsData :
-    'SELECT	kingdoms.id kingdom_id, kingdoms.name kingdomname, 1 population, locations.code location FROM kingdoms LEFT JOIN locations ON locations.kingdom_id = kingdoms.id'
-};
-
-const add = (table,params) => {
-    return save(insertQueries[table],params);
-};
-
-const get = async (selection,params) => {
-    return (await db.query(selectQueries[selection], paramsValidation(params))).results;
-};
-
-const repo = { add };
-Object.keys(selectQueries)
-  .forEach(queryName => repo['get'+queryName] = (params) => get(queryName, params));
-export const kingdomRepo = repo;
-
-
