+++ backend/src/services/index.js
@@ -1,12 +1,12 @@
 export * from './helloService';
 export * from './authService';
-export * from './userService';
+//export * from './userService';
 export * from './sessionService';
-export * from './kingdomService';
-export * from './resourceService';
+//export * from './kingdomService';
+//export * from './resourceService';
 export * from './tokenService';
-export * from './buildingService';
-export * from './class/UserService';
-export * from './class/KingdomService';
-export * from './class/ResourceService';
-export * from './class/BuildingService';
+//export * from './buildingService';
+export * from './UserService';
+export * from './KingdomService';
+export * from './ResourceService';
+export * from './BuildingService';
