+++ backend/src/services/test/UserService.test.js
@@ -67,7 +67,7 @@ test('add: valid userdata returns valid result', async () => {
     });
 });
 
-test('add: username already in use returns error 201', async () => {
+test('add: username already in use returns error 301', async () => {
     class duplicateError extends Error {
         constructor() {
             super();
@@ -83,6 +83,6 @@ test('add: username already in use returns error 201', async () => {
     try {
         const result = await user.add({userName:'username',password:'secretpassword',kingdomName:'kingdom'});
     } catch(err) {
-        expect(err).toStrictEqual( Error(201) );
+        expect(err).toStrictEqual( Error(301) );
     }
 });
